+++ app/keepers/keepers.go
@@ -29,7 +29,7 @@ import (
 	ibckeeper "github.com/cosmos/ibc-go/v7/modules/core/keeper"
 
 	callbackKeeper "github.com/archway-network/archway/x/callback/keeper"
-	custodianKeeper "github.com/archway-network/archway/x/custodian/keeper"
+	cwicaKeeper "github.com/archway-network/archway/x/cwica/keeper"
 	rewardsKeeper "github.com/archway-network/archway/x/rewards/keeper"
 	trackingKeeper "github.com/archway-network/archway/x/tracking/keeper"
 )
@@ -64,6 +64,5 @@ type ArchwayKeepers struct {
 	RewardsKeeper  rewardsKeeper.Keeper
 	CWFeesKeeper   cwfees.Keeper
 	CallbackKeeper callbackKeeper.Keeper
-
-	CustodianKeeper custodianKeeper.Keeper
+	CWICAKeeper    cwicaKeeper.Keeper
 }