+++ x/cwica/keeper/params_test.go
@@ -1,0 +1,15 @@
+package keeper_test
+
+import (
+	"github.com/archway-network/archway/x/cwica/types"
+)
+
+func (s *KeeperTestSuite) TestGetParams() {
+	ctx, k := s.chain.GetContext().WithBlockHeight(100), s.chain.GetApp().Keepers.CWICAKeeper
+	params := types.DefaultParams()
+
+	err := k.SetParams(ctx, params)
+	s.Require().NoError(err)
+
+	s.Require().EqualValues(params, k.GetParams(ctx))
+}