+++ CHANGELOG.md
@@ -38,7 +38,7 @@ Contains all the PRs that improved the code without changing the behaviors.
 - [#532](https://github.com/archway-network/archway/pull/532) - Adding ADR-009 for x/callback module
 - [#527](https://github.com/archway-network/archway/pull/527) - Add x/cwfees module.
 - [#541](https://github.com/archway-network/archway/pull/541) - Update the interchaintest framework to v7
-- [#542](https://github.com/archway-network/archway/pull/542) - Adding x/custodian module
+- [#542](https://github.com/archway-network/archway/pull/542) - Adding x/cwica module
 - [#543](https://github.com/archway-network/archway/pull/543) - Bumping sdk to v0.47.9
 
 ### Improvements